requirements.txt
opencv-python
pytesseract
reportlab
pywin32
pybidi
arabic_reshaper